clean up auth tokens and update isValidToken method
