+++ src/main/resources/authenticationTokens.json
@@ -1,5 +1,5 @@
 {
-  "authentication_tokens": [
+  "authenticationTokens": [
     "331764f0-1d84-45f5-98bb-e00f9cdec3f9",
     "201ad219-8a7c-4fea-8604-e558096d0758",
     "81eaca3e-c538-4712-873f-dd1a7a84eb8e"
